refactor: split two files over the SLOC line-cap after the #2914 fix
Rebasing the #2914 fix onto current main pushed two files over their
line-cap gate:
- crates/trusty-common/src/search_index.rs: 547 SLOC (500 prod cap)
- crates/trusty-mpm/src/core/session_launch/tests.rs: 1519 SLOC (1500
  test cap)

Both resolved via a real split (repo's established sibling-test-module
pattern), not --force-add:
- search_index.rs's `#[cfg(test)] mod tests { .. }` moved to a sibling
  `search_index_tests.rs` via `#[path = "search_index_tests.rs"] mod
  tests;` (mirrors stdio.rs/stdio_tests.rs, filter.rs/filter_tests.rs,
  etc. already in this crate). Production file is now 653 lines / well
  under 500 SLOC.
- session_launch/tests.rs's new #2914 regression test moved to a new
  sibling `tests_search_index.rs` (mirrors the existing
  tests_roster.rs/tests_scaffold_gitignore.rs split pattern in the same
  module). `EnvVarGuard` promoted to `pub(super)` so the sibling module
  can reuse it.

No test content changed, weakened, or deleted -- only relocated. Line
cap now passes clean: `bash scripts/check_line_cap.sh` -> 0 violations.
